pyproject.toml
[project]
name = "sorcererdb"
version = "0.1.0"
description = "A lightweight, caching, multi-connection SQL abstraction for Python"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Eric Harris"}]
requires-python = ">=3.8"
dependencies = [
    "mysql-connector-python"
]

[project.optional-dependencies]
dev = ["pytest", "black", "mypy"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
